Moved process_xml_hex_flower, xml2dict, and etree_to_dict to xml_functions.py. I also moved str_to_tuple to a separate function instead of within process_xml_hex_flower.
